backend/backend_main/requirements.txt
Django==4.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
django-filter==23.5
mysqlclient==2.2.0
mssqldjango==1.3
psycopg2==2.9.7
psycopg2_binary==2.9.2
asgiref==3.7.2
attrs==23.1.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.17.3
jsonschema-specifications==0.1.3
Markdown==3.4.4
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.8.9
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1